Feacher SAVE-CONFIG: Extending gamestate and incorporating saveconfig record

+++ src/main/java/oogasalad/engine/records/config/model/SaveConfigRecord.java
@@ -8,7 +8,7 @@ import java.util.List;
 public record SaveConfigRecord(
     String saveName,
     int currentLevel,
-    List<Integer> scores,
+    List<Double> scores,
     int lives,
     int highScore,
     List<Integer> levelOrder) {}
